Integrated travis, fixed build.
