+++ .travis.yml
@@ -1,0 +1,38 @@
+os: linux
+
+env:
+  global:
+    - TRAVIS_BOOST_OPTS="threading=multi link=static -sNO_BZIP2=1 -sNO_ZLIB=1 address-model=64"
+    - TRAVIS_BOOST_LIBRARIES="chrono,filesystem,program_options,system,thread,test"
+
+before_install:
+  - sudo add-apt-repository ppa:ubuntu-toolchain-r/test -y # For g++-4.9
+  - sudo apt-get update -qq
+  - sudo apt-get install -qq g++-4.9
+  - sudo update-alternatives --install /usr/bin/g++ g++ /usr/bin/g++-4.9 90
+  # Install dependencies
+  - sudo apt-get install cmake libgmp-dev libssl-dev
+  # We still need boost, let's get it.
+  - wget http://downloads.sourceforge.net/project/boost/boost/1.57.0/boost_1_57_0.tar.bz2
+  - echo "910c8c022a33ccec7f088bd65d4f14b466588dda94ba2124e78b8c57db264967  boost_1_57_0.tar.bz2" | sha256sum -c
+  - tar -xjf boost_*.tar.bz2
+  - cd boost_1_57_0
+  - ./bootstrap.sh --without-icu --with-libraries="$TRAVIS_BOOST_LIBRARIES"
+  - ./b2 -d2 -j2 -d1 $TRAVIS_BOOST_OPTS
+  - sudo ./b2 -d0 -j4 $TRAVIS_BOOST_OPTS install
+  - cd ..
+  # We still need crypto++, let's get it.
+  - mkdir cryptopp562 && cd cryptopp562
+  - wget http://www.cryptopp.com/cryptopp562.zip
+  - echo "5cbfd2fcb4a6b3aab35902e2e0f3b59d9171fee12b3fc2b363e1801dfec53574  cryptopp562.zip" | sha256sum -c
+  - unzip cryptopp562.zip
+  - make static CXXFLAGS='-DNDEBUG -g -O2 -fPIC'
+  - sudo make install
+  - cd ..
+
+script:
+  - LDFLAGS="-lrt" LIBSNARK_FLAGS="NO_PROCPS=1 NO_GTEST=1 NO_DOCS=1 STATIC=1" ./get-libsnark
+  - NO_PROCPS=1 MINDEPS=1 LINK_RT=1 make
+  - ./zerocash_pour_ppzksnark/tests/test_zerocash_pour_ppzksnark
+  - ./tests/merkleTest
+  - ./tests/zerocashTest
