+++ get-libsnark
@@ -13,8 +13,8 @@ DEPINST=`pwd -P`/$DEPINST  # remember absolute path
 
 mkdir -p $DEPSRC
 cd $DEPSRC
-
-[ ! -d libsnark ] && git clone https://github.com/scipr-lab/libsnark libsnark
+# TODO: switch back to scipr-lab/libsnark once they fix #15
+[ ! -d libsnark ] && git clone https://github.com/ebfull/libsnark libsnark
 cd libsnark
 git pull
 # TODO: submit -fPIC patch to ate-pairing
