+++ tests/zerocashTest.cpp
@@ -24,7 +24,7 @@
 #include "libzerocash/PourTransaction.h"
 #include "libzerocash/utils/util.h"
 
-#include "streams.h"
+#include "libzerocash/streams.h"
 
 using namespace std;
 using namespace libsnark;
